% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-confounders.R
\name{count_non_confounders}
\alias{count_non_confounders}
\alias{plot_non_confounders}
\title{Count and plot non-confounders}
\usage{
count_non_confounders(.df)

plot_non_confounders(.df, ..., geom = ggplot2::geom_col, sort = TRUE)
}
\arguments{
\item{.df}{A data frame, usually the result of \code{metaconfoundr()}}

\item{...}{Arguments passed to \code{geom}}

\item{geom}{The ggplot2 geom to use}

\item{sort}{Logical. Should the results be sorted?}
}
\value{
a \code{tibble} or ggplot
}
\description{
Count and plot non-confounders
}
\examples{

ipi \%>\%
  metaconfoundr() \%>\%
  plot_non_confounders(size = 3, geom = ggplot2::geom_point)

}
