% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{facet_constructs}
\alias{facet_constructs}
\title{Facet by constructs}
\usage{
facet_constructs(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}}
}
\value{
a facet component
}
\description{
A helper function to facet by constructs in `\code{\link[=mc_heatmap]{mc_heatmap()}} and
\code{\link[=mc_trafficlight]{mc_trafficlight()}}
}
\seealso{
Other plots: 
\code{\link{geom_cochrane}()},
\code{\link{mc_heatmap}()},
\code{\link{scale_fill_cochrane}()},
\code{\link{theme_mc}()}
}
\concept{plots}
