% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{label_robins}
\alias{label_robins}
\title{Label values using ROBINS approach}
\usage{
label_robins()
}
\value{
a character vector of ROBINS labels
}
\description{
\code{label_robins()} is a helper function to modify metaconfoundr labels to use
ROBINS-like labels: low risk, some concerns, high risk.
}
\examples{
mc_heatmap(metaconfoundr(ipi)) +
  ggplot2::scale_fill_ordinal(labels = label_robins())

mc_heatmap(metaconfoundr(ipi)) +
  scale_fill_cochrane(labels = label_robins())

}
