% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{launch_metaconfoundr_app}
\alias{launch_metaconfoundr_app}
\title{Launch metaconfoundr Shiny app}
\usage{
launch_metaconfoundr_app()
}
\value{
A Shiny app
}
\description{
\code{launch_metaconfoundr_app()} launches a Shiny app to create visualizations of
confounding control in meta-analyses
}
