% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mc_heatmap}
\alias{mc_heatmap}
\alias{mc_trafficlight}
\title{Plot a heatmap or traffic light plot of \code{metaconfoundr()} summaries}
\usage{
mc_heatmap(
  .df,
  legend_title = "control quality",
  sort = FALSE,
  by_group = FALSE,
  score = c("adequate", "sum", "controlled"),
  non_confounders = FALSE
)

mc_trafficlight(
  .df,
  size = 8,
  legend_title = "control quality",
  sort = FALSE,
  by_group = FALSE,
  score = c("adequate", "sum", "controlled"),
  non_confounders = FALSE
)
}
\arguments{
\item{.df}{A data frame, usually the result of \code{metaconfoundr()}}
}
\value{
a ggplot
}
\description{
\code{mc_heatmap()} and \code{mc_trafficlight()} visualize the results of
\code{metaconfoundr()}, summarizing the quality of confounder control in each
study.
}
\examples{

ipi \%>\%
  metaconfoundr() \%>\%
  mutate(variable = str_wrap(variable, 10)) \%>\%
  mc_heatmap() +
  theme_mc() +
  facet_constructs() +
  guides(x = guide_axis(n.dodge = 2))

ipi \%>\%
  metaconfoundr() \%>\%
  mc_trafficlight() +
  geom_cochrane() +
  facet_constructs() +
  scale_fill_cochrane() +
  theme_mc() +
  guides(x = guide_axis(n.dodge = 2))

}
\seealso{
Other plots: 
\code{\link{facet_constructs}()},
\code{\link{geom_cochrane}()},
\code{\link{scale_fill_cochrane}()},
\code{\link{theme_mc}()}
}
\concept{plots}
