% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaconfoundr.R
\name{mc_longer}
\alias{mc_longer}
\title{Title}
\usage{
mc_longer(
  study = contains("construct"),
  construct = contains("construct"),
  variable = matches("variable|factor"),
  control_quality = contains("control_quality"),
  is_confounder = contains("confounder"),
  study_values = mc_study_values()
)
}
\arguments{
\item{study_values}{}
}
\value{

}
\description{
Title
}
