% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaconfoundr.R
\name{mc_wider}
\alias{mc_wider}
\title{Title}
\usage{
mc_wider(
  construct = contains("construct"),
  variable = matches("variable|factor"),
  is_confounder = contains("confounder"),
  study = everything(),
  study_values = mc_study_values()
)
}
\arguments{
\item{study_values}{}
}
\value{

}
\description{
Title
}
