% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_fill_cochrane}
\alias{scale_fill_cochrane}
\alias{scale_color_cochrane}
\alias{scale_shape_cochrane}
\title{Add Cochrane-style palettes to ggplots}
\usage{
scale_fill_cochrane(...)

scale_color_cochrane(...)

scale_shape_cochrane(...)
}
\arguments{
\item{...}{Arguments passed to the underline scale function}
}
\value{
scales for ggplot
}
\description{
Add Cochrane-style palettes to ggplots
}
\seealso{
Other plots: 
\code{\link{facet_constructs}()},
\code{\link{geom_cochrane}()},
\code{\link{mc_heatmap}()},
\code{\link{theme_mc}()}
}
\concept{plots}
