% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_mc}
\alias{theme_mc}
\title{A minimal theme for metaconfoundr plots}
\usage{
theme_mc(base_size = 14)
}
\arguments{
\item{base_size}{base font size, given in pts.}
}
\value{
a ggplot theme
}
\description{
A minimal theme for metaconfoundr plots
}
\seealso{
Other plots: 
\code{\link{facet_constructs}()},
\code{\link{geom_cochrane}()},
\code{\link{mc_heatmap}()},
\code{\link{scale_fill_cochrane}()}
}
\concept{plots}
