% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mc_heatmap}
\alias{mc_heatmap}
\title{Title}
\usage{
mc_heatmap(
  .df,
  legend_title = "control quality",
  sort = FALSE,
  by_group = FALSE,
  score = c("sum", "adequate", "controlled"),
  non_confounders = FALSE
)
}
\arguments{
\item{.df}{}
}
\value{

}
\description{
Title
}
