% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaconfoundr-package.R
\docType{package}
\name{metaconfoundr-package}
\alias{metaconfoundr-package}
\alias{_PACKAGE}
\alias{metaconfoundr2}
\title{metaconfoundr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/malcolmbarrett/metaconfoundr}
  \item Report bugs at \url{https://github.com/malcolmbarrett/metaconfoundr/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{internal}
