% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaconfoundr-package.R
\docType{package}
\name{metaconfoundr-package}
\alias{metaconfoundr-package}
\alias{_PACKAGE}
\alias{metaconfoundr2}
\title{metaconfoundr: Visualize Confounder Control in Meta-Analyses}
\description{
Visualize confounder control in meta-analysis. 'metaconfoundr' is an approach to evaluating bias in studies used in meta-analyses based on the causal inference framework. Study groups create a causal diagram displaying their assumptions about the scientific question. From this, they develop a list of important confounders. Then, they evaluate whether studies controlled for these variables well. 'metaconfoundr' is a toolkit to facilitate this process and visualize the results as heatmaps, traffic light plots, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/malcolmbarrett/metaconfoundr}
  \item Report bugs at \url{https://github.com/malcolmbarrett/metaconfoundr/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

Authors:
\itemize{
  \item Julie M. Petersen \email{juliemo@bu.edu} (\href{https://orcid.org/0000-0001-7845-4545}{ORCID})
  \item Ludovic Trinquart \email{ludovic@bu.edu} (\href{https://orcid.org/0000-0002-3028-4900}{ORCID})
}

}
\keyword{internal}
